rename cached.rs to in_mem.rs (#609)
