+++ firewood/src/shale/in_mem.rs
